Move concat_map to __init__ file
